Update pyuvdata and hera_pspec versions
Avoids conflicts between the latest hera_pspec version available on PyPi at the time of this writing and the latest pyuvdata version on conda-forge.  The latest pyuvdata removes the use of e.g. future_array_shapes which modifies the shapes of the data and frequency arrays in a way that is incompatible with the latest hera_pspec on PyPi.